# docs/swagger.json
{
    "consumes": [
        "application/json",
        "application/octet-stream"
    ],
    "produces": [
        "application/json",
        "application/octet-stream"
    ],
    "swagger": "2.0",
    "info": {
        "description": "OpenTimestamps calendar server and SDK (Go), with Bitcoin anchoring.",
        "title": "OTS API",
        "termsOfService": "https://github.com/thalestmm/ots",
        "contact": {
            "name": "OTS",
            "url": "https://github.com/thalestmm/ots"
        },
        "license": {
            "name": "LGPL-3.0",
            "url": "https://www.gnu.org/licenses/lgpl-3.0.html"
        },
        "version": "0.2.0"
    },
    "host": "localhost:14788",
    "basePath": "/",
    "paths": {
        "/api/v1/health": {
            "get": {
                "description": "Reports server health; when the stamper is enabled, also checks Bitcoin RPC reachability.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_server.HealthResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/api_server.HealthResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/stamp-file": {
            "post": {
                "description": "Multipart upload (field \"file\"). The file is hashed server-side with SHA-256 and submitted to the calendar. Returns a standard detached .ots proof as bytes. Only the hash is retained.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Stamp a file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to timestamp",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detached .ots proof bytes",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/status": {
            "get": {
                "description": "Operational snapshot: pending commitments, unconfirmed anchor transactions, wallet balance, best block height.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Calendar and stamper status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_server.StatusResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/timestamps": {
            "post": {
                "description": "Convenience JSON wrapper around digest submission. Digest must be hex-encoded SHA-256 (32 bytes).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "timestamps"
                ],
                "summary": "Create timestamp (JSON)",
                "parameters": [
                    {
                        "description": "Digest to timestamp",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_server.CreateTimestampRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_server.CreateTimestampResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api_server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/upgrade": {
            "post": {
                "description": "Resolves pending attestations against this calendar and returns the (possibly) upgraded proof. complete=true once a Bitcoin attestation is present.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "timestamps"
                ],
                "summary": "Upgrade timestamp proof",
                "parameters": [
                    {
                        "description": "Digest and proof",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_server.UpgradeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_server.UpgradeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/verify": {
            "post": {
                "description": "Verifies digest binding, resolves pending attestations against this calendar, and checks Bitcoin attestations against block headers. valid=true only for cryptographically confirmed proofs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "timestamps"
                ],
                "summary": "Verify timestamp proof",
                "parameters": [
                    {
                        "description": "Digest and proof",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_server.VerifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_server.VerifyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/verify-file": {
            "post": {
                "description": "Multipart upload: field \"file\" (original content) and field \"ots\" (detached proof). Returns the full verification result.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Verify a file against its .ots proof",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Original file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Detached .ots proof",
                        "name": "ots",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_server.VerifyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/digest": {
            "post": {
                "description": "OpenTimestamps-native endpoint. Body is raw digest bytes (max 64). Returns serialized timestamp proof.",
                "consumes": [
                    "application/octet-stream"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "timestamps"
                ],
                "summary": "Submit digest for timestamping",
                "parameters": [
                    {
                        "description": "Raw digest bytes",
                        "name": "digest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Serialized OTS timestamp",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/timestamp/{commitment}": {
            "get": {
                "description": "OpenTimestamps-native upgrade endpoint. Commitment is hex-encoded bytes. The 404 body distinguishes commitments awaiting Bitcoin confirmation from unknown ones.",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "timestamps"
                ],
                "summary": "Get upgraded timestamp proof",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hex-encoded commitment",
                        "name": "commitment",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Serialized OTS timestamp",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid commitment",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Pending or not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api_server.AttestationInfoJSON": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "api_server.CreateTimestampRequest": {
            "type": "object",
            "properties": {
                "digest": {
                    "type": "string",
                    "example": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
                }
            }
        },
        "api_server.CreateTimestampResponse": {
            "type": "object",
            "properties": {
                "commitment": {
                    "type": "string"
                },
                "proof": {
                    "type": "string"
                }
            }
        },
        "api_server.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "api_server.HealthResponse": {
            "type": "object",
            "properties": {
                "bitcoin": {
                    "type": "string",
                    "example": "ok"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "api_server.StatusResponse": {
            "type": "object",
            "properties": {
                "best_block_height": {
                    "type": "integer"
                },
                "bitcoin_network": {
                    "type": "string"
                },
                "calendar_uri": {
                    "type": "string"
                },
                "last_tx_time": {
                    "type": "string"
                },
                "pending_commitments": {
                    "type": "integer"
                },
                "stamper_enabled": {
                    "type": "boolean"
                },
                "unconfirmed_txs": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                },
                "wallet_balance_btc": {
                    "type": "number"
                }
            }
        },
        "api_server.UpgradeRequest": {
            "type": "object",
            "properties": {
                "digest": {
                    "type": "string",
                    "example": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
                },
                "proof": {
                    "type": "string"
                }
            }
        },
        "api_server.UpgradeResponse": {
            "type": "object",
            "properties": {
                "complete": {
                    "type": "boolean"
                },
                "proof": {
                    "type": "string"
                },
                "upgraded": {
                    "type": "boolean"
                }
            }
        },
        "api_server.VerifyRequest": {
            "type": "object",
            "properties": {
                "digest": {
                    "type": "string",
                    "example": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
                },
                "proof": {
                    "type": "string"
                }
            }
        },
        "api_server.VerifyResponse": {
            "type": "object",
            "properties": {
                "attestations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api_server.AttestationInfoJSON"
                    }
                },
                "block_hash": {
                    "type": "string"
                },
                "block_height": {
                    "type": "integer",
                    "example": 850000
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "confirmed"
                },
                "valid": {
                    "type": "boolean"
                },
                "verified_at": {
                    "type": "string",
                    "example": "2026-06-12T14:00:00Z"
                }
            }
        }
    }
}